third_party/blink/web_tests/platform/generic/fast/dom/clone-node-default-argument-expected.txt

PASS Node.prototype.cloneNode.length is 0
PASS div.cloneNode().hasChildNodes() is false
PASS div.cloneNode().childNodes.length is 0
PASS successfullyParsed is true

TEST COMPLETE
